Add files via upload
This is the working plane controls and crash scenario. I have basically added that if the altitude of the plane is below 200 ft after 1000 iterations (100 seconds) then the plane goes into crash mode and blinks all its lights. I will after collaboration with Desmond change the altitude and iterations to what the pilot experiences in G force from the plane so that in the event of a crash the pilot will experience a G force of over 6 then the plane will go into crash mode.